Added reference deletion to all reference owners to prevent orphans. Added domain event for reference deletion. Added handlers for interface deleted and reference deleted to remove bindings before referred reference is removed.

+++ Core.DomainServices/Repositories/Reference/ReferenceRepository.cs
@@ -6,7 +6,9 @@ using Core.DomainModel.ItProject;
 using Core.DomainModel.ItSystem;
 using Core.DomainModel.ItSystemUsage;
 using Core.DomainModel.References;
+using Core.DomainModel.References.DomainEvents;
 using Core.DomainModel.Result;
+using Infrastructure.Services.DomainEvents;
 
 namespace Core.DomainServices.Repositories.Reference
 {
@@ -32,19 +34,27 @@ namespace Core.DomainServices.Repositories.Reference
         private readonly IGenericRepository<ItSystem> _systemRepository;
         private readonly IGenericRepository<ItSystemUsage> _systemUsageRepository;
         private readonly IGenericRepository<ItProject> _projectRepository;
+        private readonly IDomainEvents _domainEvents;
 
         public ReferenceRepository(
             IGenericRepository<ExternalReference> referenceRepository,
             IGenericRepository<ItContract> contractRepository,
             IGenericRepository<ItSystem> systemRepository,
             IGenericRepository<ItSystemUsage> systemUsageRepository,
-            IGenericRepository<ItProject> projectRepository)
+            IGenericRepository<ItProject> projectRepository,
+            IDomainEvents domainEvents)
         {
             _referenceRepository = referenceRepository;
             _contractRepository = contractRepository;
             _systemRepository = systemRepository;
             _systemUsageRepository = systemUsageRepository;
             _projectRepository = projectRepository;
+            _domainEvents = domainEvents;
+        }
+
+        public Maybe<ExternalReference> Get(int referenceId)
+        {
+            return _referenceRepository.GetByKey(referenceId);
         }
 
         public Maybe<IEntityWithExternalReferences> GetRootEntity(int id, ReferenceRootType rootType)
@@ -99,6 +109,11 @@ namespace Core.DomainServices.Repositories.Reference
 
         public void Delete(ExternalReference reference)
         {
+            if (reference == null)
+            {
+                throw new ArgumentNullException(nameof(reference));
+            }
+            _domainEvents.Raise(new ExternalReferenceDeleted(reference));
             _referenceRepository.Delete(reference);
             _referenceRepository.Save();
         }
